+++ tests/OpenAPITypeProvider.Tests/BasicTests.fs
@@ -5,7 +5,7 @@ open OpenAPITypeProvider
 open Newtonsoft.Json
 open System
 
-type PetStore = OpenAPIV3Provider<"Samples/PetStore.yaml">
+type PetStore = OpenAPIV3Provider<"Samples/PetStore.yaml ">
 type PetStoreJson = OpenAPIV3Provider<"Samples/PetStoreJson.json">
 
 let customDateFormat = "dd. MM. yyyy HH:mm:ss"
@@ -198,6 +198,28 @@ let ``Parses and converts simple value schema (obj array)``() =
     Assert.AreEqual(json, parsed.ToJson(Formatting.None))
     Assert.AreEqual([ob], instance.Values)
     Assert.AreEqual("Roman", instance.Values.[0].Message)
+
+[<Test>]
+let ``Parses and converts basic schema with array``() =
+    let json = """{"items":[{"name":"A","surname":"B"}]}"""
+    let ob = PetStore.Schemas.Items("A","B")
+    let instance = PetStore.Schemas.WithArrayOfObjects([ob])
+    let parsed = PetStore.Schemas.WithArrayOfObjects.Parse json
+    Assert.AreEqual(json, instance.ToJson(Formatting.None))
+    Assert.AreEqual(json, parsed.ToJson(Formatting.None))
+    Assert.AreEqual([ob], instance.Items)
+    Assert.AreEqual("B", instance.Items.[0].Surname)
+
+[<Test>]
+let ``Parses and converts basic schema with array (optional)``() =
+    let json = """{"items":[{"name":"A","surname":"B"}]}"""
+    let ob = PetStore.Schemas.Items("A","B")
+    let instance = PetStore.Schemas.WithArrayOfObjectsOptional(Some [ob])
+    let parsed = PetStore.Schemas.WithArrayOfObjectsOptional.Parse json
+    Assert.AreEqual(json, instance.ToJson(Formatting.None))
+    Assert.AreEqual(json, parsed.ToJson(Formatting.None))
+    Assert.AreEqual([ob], instance.Items.Value)
+    Assert.AreEqual("B", instance.Items.Value.[0].Surname)
 
 [<Test>]
 let ``Parses and converts simple value schema (date)``() =